# Add initial AVX512 support for dot product on Linux