@congdm 2023-03-21
Thank you for your effort! It also works on Windows and gives a little boost on my i7-11700F, from ~208 ms/token to 195 ms/token or sometimes even 185 ms/token on Alpaca7B.